requirements.txt
matplotlib
numpy
pandas
scipy
plotly
seaborn
colorama
Pillow
PyQt5
bokeh
dash
Flask==3.0.3
waitress==2.1.2
gunicorn==22.0.0
mysql-connector-python==9.0.0
SQLAlchemy==2.0.19
PyMySQL==1.0.2
greenlet==3.0.3